+++ CMimircache/cache/include/AMP.h
@@ -93,11 +93,11 @@ extern void         AMP_destroy(struct_cache* cache);
 extern void         AMP_destroy_unique(struct_cache* cache);
 
 
-struct_cache*       AMP_init(guint64 size, char data_type, int block_size, void* params);
+struct_cache*       AMP_init(guint64 size, char data_type, guint64 block_size, void* params);
 
 
 extern void         AMP_remove_element(struct_cache* cache, void* data_to_remove);
-extern gint64     AMP_get_size(struct_cache* cache);
+extern guint64    AMP_get_size(struct_cache* cache);
 
 
 #ifdef __cplusplus
