+++ CMimircache/cache/include/LRU.h
@@ -44,11 +44,11 @@ extern void     LRU_destroy(struct_cache* cache);
 extern void     LRU_destroy_unique(struct_cache* cache);
 
 
-struct_cache*   LRU_init(guint64 size, char data_type, int block_size, void* params);
+struct_cache*   LRU_init(guint64 size, char data_type, guint64 block_size, void* params);
 
 
 extern void     LRU_remove_element(struct_cache* cache, void* data_to_remove);
-extern gint64 LRU_get_size(struct_cache* cache);
+extern guint64 LRU_get_size(struct_cache* cache);
 
 
 #ifdef __cplusplus
