+++ CMimircache/cache/include/MRU.h
@@ -40,7 +40,7 @@ extern  gboolean MRU_add_element(struct_cache* cache, cache_line* cp);
 extern  void MRU_destroy(struct_cache* cache);
 extern  void MRU_destroy_unique(struct_cache* cache);
 
-struct_cache* MRU_init(guint64 size, char data_type, int block_size, void* params);
+struct_cache* MRU_init(guint64 size, char data_type, guint64 block_size, void* params);
 
 
 #ifdef __cplusplus
