+++ CMimircache/cache/include/Optimal.h
@@ -58,7 +58,7 @@ extern "C"
     extern  void optimal_destroy(struct_cache* cache);
     extern  void optimal_destroy_unique(struct_cache* cache);
     
-    struct_cache* optimal_init(guint64 size, char data_type, int block_size, void* params);
+    struct_cache* optimal_init(guint64 size, char data_type, guint64 block_size, void* params);
     
     
 #ifdef __cplusplus
