+++ CMimircache/cache/include/PG.h
@@ -70,7 +70,7 @@ typedef struct{
     uint8_t lookahead;
     char* cache_type;
     double max_meta_data;
-    uint block_size; 
+    size_t block_size;
 }PG_init_params_t;
 
 
@@ -90,11 +90,11 @@ extern void     PG_destroy(struct_cache* cache);
 extern void     PG_destroy_unique(struct_cache* cache);
 
 
-struct_cache*   PG_init(guint64 size, char data_type, int block_size, void* params);
+struct_cache*   PG_init(guint64 size, char data_type, guint64 block_size, void* params);
 
 
 extern void     PG_remove_element(struct_cache* cache, void* data_to_remove);
-extern gint64 PG_get_size(struct_cache* cache);
+extern guint64 PG_get_size(struct_cache* cache);
 
 
 #ifdef __cplusplus
