+++ CMimircache/cache/include/Random.h
@@ -41,7 +41,7 @@ extern  gboolean Random_add_element(struct_cache* cache, cache_line* cp);
 extern  void Random_destroy(struct_cache* cache);
 extern  void Random_destroy_unique(struct_cache* cache);
 
-struct_cache* Random_init(guint64 size, char data_type, int block_size, void* params);
+struct_cache* Random_init(guint64 size, char data_type, guint64 block_size, void* params);
 
 
 #ifdef __cplusplus
