+++ CMimircache/cache/include/SLRU.h
@@ -48,11 +48,11 @@ extern void     SLRU_destroy(struct_cache* cache);
 extern void     SLRU_destroy_unique(struct_cache* cache);
 
 
-struct_cache*   SLRU_init(guint64 size, char data_type, int block_size, void* params);
+struct_cache*   SLRU_init(guint64 size, char data_type, guint64 block_size, void* params);
 
 
 extern void     SLRU_remove_element(struct_cache* cache, void* data_to_remove);
-extern gint64 SLRU_get_size(struct_cache* cache);
+extern guint64 SLRU_get_size(struct_cache* cache);
 
 
 #ifdef __cplusplus
