+++ CMimircache/cache/include/Score.h
@@ -68,7 +68,7 @@ extern  gboolean Score_add_element(struct_cache* cache, cache_line* cp);
 extern  void Score_destroy(struct_cache* cache);
 extern  void Score_destroy_unique(struct_cache* cache);
 
-struct_cache* Score_init(guint64 size, char data_type, int block_size, void* params);
+struct_cache* Score_init(guint64 size, char data_type, guint64 block_size, void* params);
 
 
 #ifdef __cplusplus
